+++ anyblok/model.py
@@ -19,6 +19,7 @@ from anyblok.common import TypeList, apply_cache
 from copy import deepcopy
 from sqlalchemy.ext.declarative import declared_attr
 from .mapper import ModelAttribute
+from sqlalchemy import ForeignKeyConstraint
 
 
 class ModelException(Exception):
@@ -405,7 +406,15 @@ class Model:
         if table_args:
             def define_table_args(cls_):
                 if cls_.__registry_name__ == namespace:
-                    return super(base, cls_).define_table_args() + table_args
+                    res = super(base, cls_).define_table_args()
+                    fks = [x.name for x in res
+                           if isinstance(x, ForeignKeyConstraint)]
+
+                    t_args = [x for x in table_args
+                              if (not isinstance(x, ForeignKeyConstraint) or
+                                  x.name not in fks)]
+
+                    return res + tuple(t_args)
 
                 return ()